✨ feat: Refactor code inline node handling and restore getCodeInlineNode function
