+++ src/plugins/code/command/index.ts
@@ -5,35 +5,14 @@ import {
   $isRangeSelection,
   COMMAND_PRIORITY_EDITOR,
   LexicalEditor,
-  LexicalNode,
   createCommand,
 } from 'lexical';
 
-import { $createCursorNode, $isCursorNode } from '@/plugins/common';
+import { $createCursorNode } from '@/plugins/common';
 
-import { $createCodeNode, $isCodeInlineNode } from '../node/code';
+import { $createCodeNode, $isCodeInlineNode, getCodeInlineNode } from '../node/code';
 
 export const INSERT_CODEINLINE_COMMAND = createCommand<undefined>('INSERT_CODEINLINE_COMMAND');
-
-function getCodeInlineNode(node: LexicalNode) {
-  if ($isCursorNode(node)) {
-    const parent = node.getParent();
-    if ($isCodeInlineNode(parent)) {
-      return parent;
-    }
-    if ($isCodeInlineNode(node.getNextSibling())) {
-      return node.getNextSibling();
-    }
-    if ($isCodeInlineNode(node.getPreviousSibling())) {
-      return node.getPreviousSibling();
-    }
-    return null;
-  }
-  if ($isCodeInlineNode(node.getParent())) {
-    return node.getParent();
-  }
-  return null;
-}
 
 export function registerCodeInlineCommand(editor: LexicalEditor) {
   return editor.registerCommand(
